refactor: pipelines folder architecture
