+++ src/etl/EtlService.ts
@@ -1,10 +1,10 @@
 import { errors } from '@elastic/elasticsearch'
 import fs from 'fs'
 
-import { IngestPipeline } from './ingest-pipelines/IngestPipeline'
-import { IngestPipelineCtis } from './ingest-pipelines/IngestPipelineCtis'
-import { IngestPipelineDmDmdiv } from './ingest-pipelines/IngestPipelineDmDmdiv'
-import { IngestPipelineJarde } from './ingest-pipelines/IngestPipelineJarde'
+import { IngestPipeline } from './pipelines/ingest/IngestPipeline'
+import { IngestPipelineCtis } from './pipelines/ingest/IngestPipelineCtis'
+import { IngestPipelineDmDmdiv } from './pipelines/ingest/IngestPipelineDmDmdiv'
+import { IngestPipelineJarde } from './pipelines/ingest/IngestPipelineJarde'
 import { S3Service } from './s3/S3Service'
 import { elasticsearchIndexMapping } from '../shared/elasticsearch/elasticsearchIndexMapping'
 import { ElasticsearchService } from '../shared/elasticsearch/ElasticsearchService'
