+++ src/etl/pipelines/ingest/IngestPipelineCtis.ts
@@ -1,8 +1,8 @@
 import { IngestPipeline } from './IngestPipeline'
-import { ResearchStudyModel } from '../../shared/models/domain-resources/ResearchStudyModel'
-import { EclaireDto } from '../dto/EclaireDto'
-import { RiphCtisDto } from '../dto/RiphCtisDto'
-import { ResearchStudyModelFactory } from '../factory/ResearchStudyModelFactory'
+import { ResearchStudyModel } from '../../../shared/models/domain-resources/ResearchStudyModel'
+import { EclaireDto } from '../../dto/EclaireDto'
+import { RiphCtisDto } from '../../dto/RiphCtisDto'
+import { ResearchStudyModelFactory } from '../../factory/ResearchStudyModelFactory'
 
 export class IngestPipelineCtis extends IngestPipeline {
   readonly type = 'ctis'
